pyproject.toml
[project]
name = "sign-pdf"
version = "0.1.0"
description = "PDF signature/signing utility project. Add signatures to PDFs easily."
readme = "README.md"
authors = [{ name = "Lone Striker" }]
requires-python = ">=3.9"
dependencies = [
    "flask>=2.0",
    "pillow>=8.0",
    # "pymupdf>=1.18.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# [tool.hatch.build.targets.sdist]
# include = [
#   "sign_pdf",
#   "web",
# ]
# exclude = [
# ]

[project.urls]
"Source code" = "https://github.com/lonestriker/sign-pdf"

[project.scripts]
sign-pdf = "sign_pdf.app:run_app"
